Fix compile error and cleanup warnings on clang
Refactored block struct to its own file to avoid forward declaration
error in c++ standard, gcc does not enforce this for some reason but
clang does.

Added util.hpp to testing to silence unused variable warnings.

Cleanup various trivial compiler warnings (extra braces etc.)

Now compiles with clang 4.0.1 and all 208 tests pass.
