+++ alb/allocator_base.hpp
@@ -19,76 +19,6 @@
 
 namespace alb {
   inline namespace v_100 {
-    /**
-     * The value type to describe a memory block and it's length
-     * \ingroup group_allocators
-     */
-    struct block {
-      block() noexcept
-        : ptr(nullptr)
-        , length(0)
-      {
-      }
-
-      block(void *ptr, size_t length) noexcept
-        : ptr(ptr)
-        , length(length)
-      {
-      }
-
-      block(block &&x) noexcept
-      {
-        *this = std::move(x);
-      }
-
-      block &operator=(block &&x) noexcept
-      {
-        ptr = x.ptr;
-        length = x.length;
-        x.reset();
-        return *this;
-      }
-
-      block &operator=(const block &x) noexcept = default;
-      block(const block &x) noexcept = default;
-
-      /**
-       * During destruction of any of this instance, the described memory
-       * is not freed!
-       */
-      ~block()
-      {
-      }
-
-      /**
-       * Clears the block
-       */
-      void reset() noexcept
-      {
-        ptr = nullptr;
-        length = 0;
-      }
-
-      /**
-       * Bool operator to make the Allocator code better readable
-       */
-      explicit operator bool() const
-      {
-        return length != 0;
-      }
-
-      bool operator==(const block &rhs) const
-      {
-        return ptr == rhs.ptr && length == rhs.length;
-      }
-
-      /// This points to the start address of the described memory block
-      void *ptr;
-
-      /// This describes the length of the reserved bytes.
-      size_t length;
-    };
-
 
     template< class T, class Allocator, class... Args>
     std::shared_ptr<T> make_shared(const Allocator& alloc, Args&&... args)
