+++ alb/memory_corruption_detector.hpp
