+++ test/util.hpp
@@ -1,0 +1,8 @@
+#ifndef UTIL_HPP
+#define UTIL_HPP
+
+template<typename T>
+void ignore_unused(T&&)
+{}
+
+#endif
